+++ spec/views/analyses/index.html.erb_spec.rb
@@ -2,23 +2,33 @@
 
 require 'rails_helper'
 
-describe 'analyses/manage', type: :view do
+describe 'analyses/index', type: :view do
   include AnalysisViewHelper
 
   before(:all) do
     Kaminari.config.default_per_page = AnalysisViewHelper::DEFAULT_PER_PAGE
-    @analysis = Analysis.new
-    @analysis.build_parameter
   end
 
   before do
-    render template: 'analyses/manage', layout: 'layouts/application'
+    render template: 'analyses/index', layout: 'layouts/application'
     @html ||= Nokogiri.parse(response)
+  end
+
+  context '検索フォームに条件が入力されている場合' do
+    attribute = {num_data: 1, parameter: {max_depth: 1, max_features: 'log2'}}
+    include_context 'トランザクション作成'
+    include_context '分析ジョブを作成する'
+    include_context '分析フォームオブジェクトを作成する', attribute
+    include_context 'HTML初期化'
+    it_behaves_like '分析画面共通テスト'
+    it_behaves_like 'ページングボタンが表示されていないこと'
+    it_behaves_like '分析ジョブの状態が正しいこと', '実行待ち'
   end
 
   context '実行待ちの場合' do
     include_context 'トランザクション作成'
     include_context '分析ジョブを作成する'
+    include_context '分析フォームオブジェクトを作成する'
     include_context 'HTML初期化'
     it_behaves_like '分析画面共通テスト'
     it_behaves_like 'ページングボタンが表示されていないこと'
@@ -29,6 +39,7 @@ describe 'analyses/manage', type: :view do
     update_attribute = {state: 'processing', performed_at: Time.zone.now}
     include_context 'トランザクション作成'
     include_context '分析ジョブを作成する', update_attribute: update_attribute
+    include_context '分析フォームオブジェクトを作成する'
     include_context 'HTML初期化'
     it_behaves_like '分析画面共通テスト'
     it_behaves_like 'ページングボタンが表示されていないこと'
@@ -39,6 +50,7 @@ describe 'analyses/manage', type: :view do
     update_attribute = {num_entry: 10, state: 'processing', performed_at: Time.zone.now}
     include_context 'トランザクション作成'
     include_context '分析ジョブを作成する', update_attribute: update_attribute
+    include_context '分析フォームオブジェクトを作成する'
     include_context 'HTML初期化'
     it_behaves_like '分析画面共通テスト'
     it_behaves_like 'ページングボタンが表示されていないこと'
@@ -49,6 +61,7 @@ describe 'analyses/manage', type: :view do
     update_attribute = {state: 'completed', performed_at: Time.zone.now}
     include_context 'トランザクション作成'
     include_context '分析ジョブを作成する', update_attribute: update_attribute
+    include_context '分析フォームオブジェクトを作成する'
     include_context 'HTML初期化'
     it_behaves_like '分析画面共通テスト'
     it_behaves_like 'ページングボタンが表示されていないこと'
@@ -59,6 +72,7 @@ describe 'analyses/manage', type: :view do
     update_attribute = {state: 'error', performed_at: Time.zone.now}
     include_context 'トランザクション作成'
     include_context '分析ジョブを作成する', update_attribute: update_attribute
+    include_context '分析フォームオブジェクトを作成する'
     include_context 'HTML初期化'
     it_behaves_like '分析画面共通テスト'
     it_behaves_like 'ページングボタンが表示されていないこと'
@@ -69,6 +83,7 @@ describe 'analyses/manage', type: :view do
   context "分析ジョブ情報が#{total}件の場合" do
     include_context 'トランザクション作成'
     include_context '分析ジョブを作成する', total: total
+    include_context '分析フォームオブジェクトを作成する'
     include_context 'HTML初期化'
     it_behaves_like '分析画面共通テスト', expected: {total: total}
     it_behaves_like 'ページングボタンが表示されていること'